sidecar/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
edge-tts>=6.1.9
requests>=2.31.0
pydub>=0.25.1
python-dotenv>=1.0.0
python-pptx>=1.0.0
duckduckgo-search>=8.1.0
Pillow>=10.0.0
psycopg[binary]>=3.1.0; platform_system == "Windows"
psycopg>=3.1.0; platform_system != "Windows"
aiohttp>=3.9.0
comtypes>=1.4.0; platform_system == "Windows"

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Vector database & embeddings
chromadb>=0.4.0
